+++ src/components/pages/apps/genes/gsea/ext-gsea/ext-gsea-provider.tsx
@@ -133,7 +133,7 @@ export function ExtGseaProvider({
       addAxes([
         {
           plotId: result.id,
-          groupId: 'ext-gsea',
+          groupId: 'es',
           axisIds: ['x'],
           axes: {
             x: xax,